requirements.txt
aniso8601==9.0.1
api==0.0.7
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
bootstrap4==0.1.0
bootstraps==1.0.1
Brlapi==0.6.6
certifi==2022.5.18.1
cffi==1.15.1
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.3.7
command-not-found==0.3
constantly==15.1.0
cryptography==37.0.4
cupshelpers==1.0
daphne==3.0.2
dataclasses==0.8
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.15
django-autoslug==1.9.8
django-ckeditor==6.3.2
django-cors-headers==3.10.1
django-filter==21.1
django-js-asset==2.0.0
django-phonenumber-field==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.2.1
Faker==13.13.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.9.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.3
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.0.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
mysqlclient==2.1.1
netifaces==0.10.4
nose==1.3.7
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
packaging==21.3
pandas==1.1.5
pbr==5.10.0
pexpect==4.2.1
phonenumbers==8.12.52
Pillow==5.1.0
pluggy==1.0.0
protobuf==3.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyJWT==2.4.0
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.1.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.0
pyrsistent==0.18.0
pytest==7.0.1
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.20.0
pytz==2022.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.27.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.13.2
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
style==1.1.0
system-service==0.3
systemd-python==234
testresources==2.0.1
tomli==1.2.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update==0.0.1
urllib3==1.26.9
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==2.0.3
xkit==0.0.0
zipp==3.6.0
zope.interface==5.4.0
